+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added 
+ - Pylace's Engine now implements deepcopy.
+
+
 ## [python-0.3.1] - 2023-08-28
 
 ### Fixed